requirements.txt
aiogram==2.23.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
netifaces==0.11.0
openai==0.27.0
peewee==3.15.4
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
python-nmap==0.7.1
python-telegram-bot-calendar==1.0.5
pytz==2023.3
requests==2.28.2
scapy==2.5.0
six==1.16.0
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
